+++ code/tp51/config/app.php
@@ -51,7 +51,7 @@ return [
     // 默认模块名
     'default_module'         => 'frontend',
     // 禁止访问模块
-    'deny_module_list'       => ['common', 'module'],
+    'deny_module_list'       => ['common', 'module', 'core', 'backups'],
     // 默认控制器名
     'default_controller'     => 'Index',
     // 默认操作名