+++ code/tp51/local/common.php
@@ -1,3 +1,4 @@
 <?php
 
+// 注册独立模块
 require_once __DIR__ . '/modules.php';