+++ code/tp51/thoughts/include/interface.php
@@ -1,0 +1,12 @@
+<?php
+namespace think;
+
+if (interface_exists('Psr\Log\LoggerInterface')) {
+    interface LoggerInterface extends \Psr\Log\LoggerInterface
+    {
+    }
+} else {
+    interface LoggerInterface
+    {
+    }
+}