+++ code/tp51/thoughts/module/doctrine/functions.php
@@ -8,3 +8,7 @@ function getEntity($entity) {
     $class = '\\YRS\\Entity\\' . $entity;
     return class_exists($class) ? new $class() : false;
 }
+
+function getClass($entity) {
+    return '\\YRS\\Entity\\' . $entity;
+}