+++ code/tp51/thoughts/module/helper/functions/assets.php
@@ -1,0 +1,9 @@
+<?php
+
+function script($src = '') {
+    return '<script src="' . $src . '" type="text/javascript"></script>';
+}
+
+function style($href = '') {
+    return '<link href="' . $href . '" rel="stylesheet">';
+}