+++ code/tp51/thoughts/system/console/Plugin.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace YRS\Composer;
+
+use Composer\Composer;
+use Composer\Installer;
+use Composer\IO\IOInterface;
+use Composer\Plugin\PluginInterface;
+
+class Plugin implements PluginInterface
+{
+    public function activate(Composer $composer, IOInterface $io)
+    {
+        $manager = $composer->getInstallationManager();
+        $manager->addInstaller(new RestoreThinkPHP($io, $composer));
+    }
+}