+++ code/tp51/thoughts/system/feature/ModulerLauncher.php
@@ -17,15 +17,13 @@ class ModulerLauncher extends Moduler
         $default = $this->default;
         if (is_array($this->modules) && !count($this->modules)) return false;
         foreach ($this->modules as $name => $class) {
-            if (!isset($default[$name])) {
-                new $class();
-            } elseif ($default[$name] instanceof \Closure) {
-                if ($default[$name]()) new $class();
-            } elseif (is_bool($default[$name])) {
-                if ($default[$name]) new $class();
-            } else {
-                // TODO: do nothing
-            }
+            if (!class_exists($class) || !isset($default[$name])) continue;
+
+            $reflection = new \ReflectionClass($class);
+            $case0 = $default[$name] instanceof \Closure && $default[$name]();
+            $case1 = is_bool($default[$name]) && $default[$name];
+
+            if ($case0 || $case1) $reflection->newInstance();
         }
         // TODO: no return
     }